+++ CHANGELOG.md
@@ -1,8 +1,20 @@
+# v1.7.28
+## 01/24/2022
+
+1. [](#new)
+    * Added links and modules support to `HtmlBlock` class
+    * Added module support for twig script tag: `{% script module 'theme://js/module.mjs' %}`
+    * Added twig tag for links: `{% link icon 'theme://images/favicon.png' priority: 20 with { type: 'image/png' } %}`
+    * Added `HtmlBlock` support for `{% style %}`, `{% script %}` and `{% link %}` tags
+    * Support for page-level `redirect_default_route` frontmatter header override
+3. [](#bugfix)
+    * Fixed XSS check not detecting escaped `&#58`
+
 # v1.7.27.1
 ## 01/12/2022
 
 3. [](#bugfix)
-   * Fixed a typo in CSS Asset pipeline that was erroneously joining files with `;`  
+   * Fixed a typo in CSS Asset pipeline that was erroneously joining files with `;`
 
 # v1.7.27
 ## 01/12/2022